refactor directory structure and filenames' convention
