+++ src/assets/expo.symbol.white.png
